fix: follow up reviews
